Remove photos hardcoded data, move backgrounds.js out of components folder
