--- components/backgrounds.js
@@ -1,77 +1,0 @@
-import black1 from '../public/black1.jpg';
-import black2 from '../public/black2.jpg';
-
-  export const background = [
-    {
-        title: "Photo title1",
-        src: black1,
-        width: 3,
-        height: 4
-    },
-    {
-        title: "Photo title2",
-        src: black1,
-        width: 3,
-        height: 4
-    },
-    {
-        title: "Photo title3",
-        src: black1,
-        width: 3,
-        height: 4
-    },
-    {
-        title: "Photo title11",
-        src: black1,
-        width: 3,
-        height: 4
-    } ,
-    {
-        title: "Photo title4",
-        src: black1,
-        width: 3,
-        height: 4
-    },
-    {
-        title: "Photo title13",
-        src: black2,
-        width: 4,
-        height: 3
-    },
-    {
-        title: "Photo title14",
-        src: black1,
-        width: 3,
-        height: 4
-    } ,
-    {
-        title: "Photo title52",
-        src: black1,
-        width: 3,
-        height: 4
-    } ,
-    {
-        title: "Photo title63",
-        src: black1,
-        width: 3,
-        height: 4
-    },
-    {
-        title: "Photo title15",
-        src: black2,
-        width: 4,
-        height: 3
-    }  ,
-    {
-        title: "Photo title74",
-        src: black1,
-        width: 3,
-        height: 4
-    } ,
-    {
-        title: "Photo title71",
-        src: black1,
-        width: 3,
-        height: 4
-    } 
-  ];
