--- components/photos.js
@@ -1,125 +1,0 @@
-import myImg1 from '../public/horse.jpg';
-import myImg2 from '../public/food.jpg';
-import myImg3 from '../public/alaska.jpg';
-import myImg4 from '../public/dog.jpg';
-import myImg5 from '../public/water.jpg';
-import myImg6 from '../public/desert.jpg';
-import myImg7 from '../public/palm.jpg';
-import myImg11 from '../public/11.jpeg';
-import myImg12 from '../public/12.jpeg';
-import myImg13 from '../public/13.jpeg';
-import myImg14 from '../public/14.jpeg';
-import black1 from '../public/black1.jpg';
-import black2 from '../public/black2.jpg';
-
-export const photos = [
-    {
-        caption: "Photo title1",
-        background: black1,
-        src: myImg1,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 3,
-        height: 4
-    },
-    {
-        caption: "Photo title2",
-        background: black1,
-        src: myImg2,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 3,
-        height: 4
-    },
-    {
-        caption: "Photo title3",
-        background: black1,
-        src: myImg3,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 4,
-        height: 3
-    },
-    {
-        caption: "Photo title11",
-        background: black1,
-        src: myImg11,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 3,
-        height: 4
-    } ,
-    {
-        title: "Photo title4",
-        background: black1,
-        src: myImg5,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 4,
-        height: 3
-    },
-    {
-        caption: "Photo title13",
-        background: black2,
-        src: myImg13,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 4,
-        height: 3
-    },
-    {
-        caption: "Photo title14",
-        background: black1,
-        src: myImg14,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 3,
-        height: 4
-    } ,
-    {
-        caption: "Photo title52",
-        background: black1,
-        src: myImg4,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 3,
-        height: 4
-    } ,
-    {
-        caption: "Photo title63",
-        background: black1,
-        src: myImg6,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 4,
-        height: 3
-    },
-    {
-        caption: "Photo title15",
-        background: black2,
-        src: myImg12,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 4,
-        height: 3
-    }  ,
-    {
-        caption: "Photo title74",
-        background: black1,
-        src: myImg7,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 3,
-        height: 4
-    } ,
-    {
-        caption: "Photo title71",
-        background: black1,
-        src: myImg2,
-        txt : "Some text to be displayed...", 
-        label: ["", "", ""],
-        width: 3,
-        height: 4
-    } 
-  ];
-
